1) move encryptor to utils
2) index jwt token
